refactor: enhance model structure and output type definitions
- Introduced a new `types.py` file for each model family (Chai-1, Boltz-2, ESM) to define lightweight output types, enabling dependency isolation and cleaner imports.
- Updated model classes to utilize the new output types, improving clarity and maintainability.
- Refactored import paths to streamline the codebase and enhance type checking.
- Removed redundant code and comments to optimize performance and readability.
